src/win-arena-container/client/requirements.txt
numpy~=1.24.3
Pillow~=10.1.0
fabric
gymnasium~=0.28.1
requests~=2.31.0
transformers~=4.36.0
torch~=2.1.1
accelerate
opencv-python~=4.8.1.78
matplotlib~=3.7.4
pynput~=1.7.6
pyautogui~=0.9.54
psutil~=5.9.6
tqdm~=4.65.0
pandas~=2.0.3
flask~=3.0.0
requests-toolbelt~=1.0.0
lxml
cssselect
xmltodict
openpyxl
python-docx
python-pptx
pypdf
PyGetWindow
rapidfuzz
pyacoustid
opencv-python
ImageHash
scikit-image
librosa
pymupdf
chardet
playwright
backoff
formulas
pydrive
fastdtw
odfpy
openai
func-timeout
beautifulsoup4
dashscope
google-generativeai
PyYaml
mutagen
easyocr
borb
pypdf2
pdfplumber
wandb
wrapt_timeout_decorator
gdown
tiktoken
groq
torchvision
addict
yapf
pycocotools
timm
pytesseract
tenacity
LnkParse3
anytree
supervision
ultralytics
py7zr
